refactor: switch to oop
